🔥 chore(IModuleAttendance.cs): remove commented out code The commit removes commented out code that is no longer needed. This improves the readability of the code and reduces clutter.
🔥 refactor(IModuleAttendance.cs): remove commented out code for MTReason and MTLocation
The code contained commented out code for MTReason and MTLocation which is no longer needed. This commit removes the commented out code to improve code readability and maintainability.

✨ feat(IModuleAttendance.cs): add TRTimeot region
This commit adds a new region for TRTimeot in the IModuleAttendance.cs file. It includes three new methods for getting, managing, and deleting TRTimeot.