🚀 feat(DataModuleAttendance.cs): add InputTRTimeshift class The InputTRTimeshift class is added to the DataModuleAttendance.cs file. This class contains properties that represent the input data for a timeshift. These properties include device_name, ip, username, company_code, worker_code, worker_name, timeshift_id, timeshift_doc, timeshift_workdate, timeshift_old, timeshift_new, timeshift_note, reason_code, status, modified_by, modified_date, timeshift_fromdate, timeshift_todate, flag, and timeshift_data. This class is used to represent the input data for a timeshift.
🚀 feat(IModuleAttendance.cs): add TRTimeshift region with 3 methods
This commit adds a new region called TRTimeshift to the IModuleAttendance.cs file. The region contains three new methods: getTRTimeshiftList, doManageTRTimeshift, and doDeleteTRTimeshift. These methods are used to manage TRTimeshift data and are implemented as WebInvoke methods with JSON request and response formats.

✨ feat(ModuleAttendance.svc.cs): add support for TRTimeshift module
This commit adds support for the TRTimeshift module in the HRFocusWCF application. The `getTRTimeshiftList` function retrieves a list of timeshift data based on the input parameters. The `doManageTRTimeshift` function manages the timeshift data by inserting new data or updating existing data. The `doDeleteTRTimeshift` function deletes the timeshift data. These functions are now available in the application.